use explicit cast to double when promoting float lookup value from coulomb table

+++ src/DIELECTRIC/pair_lj_long_coul_long_dielectric.cpp
@@ -183,7 +183,7 @@ void PairLJLongCoulLongDielectric::compute(int eflag, int vflag)
           rsq_lookup.f = rsq;
           itable = rsq_lookup.i & ncoulmask;
           itable >>= ncoulshiftbits;
-          fraction = (rsq_lookup.f - rtable[itable]) * drtable[itable];
+          fraction = (double) (rsq_lookup.f - rtable[itable]) * drtable[itable];
           table = ftable[itable] + fraction * dftable[itable];
           force_coul = qtmp * q[j] * table;
           efield_i = q[j] * table / qqrd2e;
